Migration to RA2 (#692)
* migrate fluid canner recipes

* migrate forge hammer recipes

* migrate DTPF recipes (1/?)

* migrate DTPF recipes (2/2)

* migrate laser engraver recipes

* migrate chemical reactor recipes (1/2)

* migrate cutting machine recipes (1/?)

* spotlessApply (#693)


* migrate extractor recipes

* migrate chemical reactor recipes (2/2)

* spotless apply

* migrate electrolyser recipes

* Revert "migrate extractor recipes"

This reverts commit 89aa54c7

* Revert "migrate cutting machine recipes (1/?)"

This reverts commit a8322d59

* Revert "migrate chemical reactor recipes (1/2)"

This reverts commit b5e26330

* Revert "migrate laser engraver recipes"

This reverts commit 3c19f2316f95f29be9fb4f383030541899c90d29.

* Revert "migrate electrolyser recipes"

This reverts commit 93189311579e4ed68f87f121938d2086ce40035c.

* migrate extractor recipes

* migrate laser engraver recipes

* spotless apply

* migrate electrolyzer recipes

* migrate centrifuge recipes

* migrate cutting machine recipes

* migrate chemical reactor recipes (1/2)

* migrate chemical reactor recipes (2/2)

* migrate assembler recipes from scripts

* migrate circuit assembler recipes from scripts

* migrate cutting machine recipes from scripts

* convert fuel recipes to RA2

* migrate cracking recipes

* migrate missed RA1 recipes

* attempt to fix centrifuge recipes #1

* spotless

* one more centrifuge fix

* remove unnecessary 100% bloat

* fix ammonia

* fix output gas in 6 recipes

* fix ae2fc recipe

* coding style

* fix

* spotless

* fix recipe map for superheavy radox

* readd missing slime extraction recipes

* add nooptimize and remove unnecessary comments

* fix biotems

* update deps

* some more nooptimize

---------

+++ src/main/java/com/dreammaster/gthandler/recipes/CircuitAssemblerRecipes.java
@@ -229,8 +229,8 @@ public class CircuitAssemblerRecipes implements Runnable {
                             GT_OreDictUnificator.get(OrePrefixes.wireFine, Materials.NiobiumTitanium, 16),
                             ELEMENT.STANDALONE.CHRONOMATIC_GLASS.getBolt(8))
                     .itemOutputs(ItemList.Circuit_Bioprocessor.get(1L)).fluidInputs(new FluidStack(solderUEV, 144))
-                    .noFluidOutputs().requiresCleanRoom().duration(3 * SECONDS + 15 * TICKS)
-                    .eut((int) TierEU.RECIPE_UEV).addTo(sCircuitAssemblerRecipes);
+                    .noFluidOutputs().requiresCleanRoom().duration(3 * SECONDS + 15 * TICKS).eut(TierEU.RECIPE_UEV)
+                    .addTo(sCircuitAssemblerRecipes);
         }
 
         GT_Values.RA.stdBuilder()
